- removed bootstrap and node-sass dependency - bundled all app related files inside app directory - renamed files
