+++ src/app/app.test.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import { render } from "@testing-library/react";
+
+import App from "./app";
+
+test("renders learn react link", () => {
+  const { getByText } = render(<App />);
+  const linkElement = getByText(/learn react/i);
+  expect(linkElement).toBeInTheDocument();
+});
