+++ src/app/app.types.ts
@@ -1,0 +1,4 @@
+export const appTypes = {
+  SET_DARK_THEME: "SET_DARK_THEME",
+  SET_LIGHT_THEME: "SET_LIGHT_THEME",
+};
