--- src/layouts/MasterLayout.tsx
@@ -1,10 +1,0 @@
-import React from "react";
-import styled from "styled-components";
-interface Props {}
-const MasterLayout: React.FC<Props> = () => {
-  return <Wrapper>Hello everyone</Wrapper>;
-};
-export default MasterLayout;
-const Wrapper = styled.div`
-  width: 100%;
-`;
