+++ src/layouts/master-layout.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+interface Props {}
+const MasterLayout: React.FC<Props> = () => {
+  return <div>Hello everyone</div>;
+};
+export default MasterLayout;
