+++ src/screens/Toggle.tsx
@@ -1,6 +1,5 @@
 import React from "react";
 import { Theme } from "../enum/theme";
-import { ToggleContainer } from "./toggle.styled";
 import { ReactComponent as SunIcon } from "../assets/icons/sun.svg";
 import { ReactComponent as MoonIcon } from "../assets/icons/moon.svg";
 
@@ -8,13 +7,12 @@ interface Props {
   theme: Theme;
   toggleTheme: () => void;
 }
-const Toggle: React.FC<Props> = ({ theme, toggleTheme }) => {
-  const isLight = theme === Theme.LIGHT;
+const Toggle: React.FC<Props> = ({ toggleTheme }) => {
   return (
-    <ToggleContainer onClick={toggleTheme} lightTheme={isLight}>
+    <div onClick={toggleTheme} className="toggle-container">
       <SunIcon />
       <MoonIcon />
-    </ToggleContainer>
+    </div>
   );
 };
