+++ src/styles/app.scss
@@ -1,6 +1,5 @@
 @import "normalize";
-@import "node_modules/bootstrap/scss/bootstrap.scss";
 @import "variables";
 @import "color-theme";
 @import "layout";
-@import 'bootstrap.min.css';
+@import "bootstrap.min.css";
